Move consensus state to separate module
