+++ x/ibc/src/ics02_client/types/mod.rs
@@ -1,2 +1,3 @@
 pub mod client_state;
+pub mod consensus_state;
 pub mod query;
